relocate base-component to base subdir and update dependants' imports
