+++ src/components/base/base-flip-card.js
@@ -1,6 +1,6 @@
 import { LitElement, css, customElement, html, property } from 'lit-element';
 
-import BaseComponent from '../base-component';
+import BaseComponent from './base-component';
 
 @customElement('base-flip-card')
 class BaseFlipCard extends BaseComponent {
